[refac] Move 3d rendering in own component
